% Auto-generated: do not edit by hand
\name{videoPreview}

\alias{videoPreview}

\title{VideoPreview component}

\description{

}

\usage{
videoPreview(id=NULL, datasetStartTime=NULL, isPlaying=NULL,
playbackRate=NULL, playheadTime=NULL, showControls=NULL,
style=NULL, timeOffset=NULL, videoMetadata=NULL,
videoSrc=NULL)
}

\arguments{
\item{id}{Character. }

\item{datasetStartTime}{Numeric. }

\item{isPlaying}{Logical. }

\item{playbackRate}{Numeric. }

\item{playheadTime}{Numeric. }

\item{showControls}{Logical. }

\item{style}{Named list. }

\item{timeOffset}{Numeric. }

\item{videoMetadata}{Named list. }

\item{videoSrc}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

