% Auto-generated: do not edit by hand
\name{videoPreview}

\alias{videoPreview}

\title{VideoPreview component}

\description{

}

\usage{
videoPreview(id=NULL, currentTime=NULL, endTime=NULL, isPlaying=NULL,
playheadTime=NULL, startTime=NULL, style=NULL, videoSrc=NULL)
}

\arguments{
\item{id}{Character. }

\item{currentTime}{Numeric. }

\item{endTime}{Numeric. }

\item{isPlaying}{Logical. }

\item{playheadTime}{Numeric. }

\item{startTime}{Numeric. }

\item{style}{Named list. }

\item{videoSrc}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

